# app/api/agents/route.ts
import { NextResponse } from "next/server"
import { createSystemEventResponse } from "@/lib/events/event-stream"
import { listRegisteredAgents, registerAgent } from "@/lib/agent-registry"
import { agentRegistrationSchema } from "@/lib/openapi/schemas"

export const dynamic = "force-dynamic"

export async function GET(req: Request) {
  const url = new URL(req.url)

  if (url.searchParams.get("stream") === "1") {
    return createSystemEventResponse()
  }

  const agents = listRegisteredAgents({
    district: url.searchParams.get("district") ?? undefined,
    status: url.searchParams.get("status") ?? undefined,
    skill: url.searchParams.get("skill") ?? undefined,
    capability: url.searchParams.get("capability") ?? undefined,
  })

  return NextResponse.json(
    { ok: true, agents },
    { headers: { "Cache-Control": "no-store" } },
  )
}

export async function POST(req: Request) {
  try {
    const parsed = agentRegistrationSchema.safeParse(await req.json())
    if (!parsed.success) {
      return NextResponse.json(
        {
          ok: false,
          error: "Invalid agent registration",
          ...(process.env.NODE_ENV === "development"
            ? { details: parsed.error.flatten() }
            : {}),
        },
        { status: 400 },
      )
    }

    const agent = registerAgent(parsed.data)
    return NextResponse.json(
      { ok: true, agent },
      { status: 201, headers: { "Cache-Control": "no-store" } },
    )
  } catch (error) {
    return NextResponse.json(
      { ok: false, error: error instanceof Error ? error.message : "Failed registering agent" },
      { status: 400 },
    )
  }
}